refactor: 데이터 레이어 패키지 구조 재구성 및 UI 모델 위치 변경
