+++ app/src/main/java/woowacourse/shopping/data/local/dao/RecentProductDao.kt
@@ -1,10 +1,10 @@
-package woowacourse.shopping.local.dao
+package woowacourse.shopping.data.local.dao
 
 import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
-import woowacourse.shopping.local.entity.RecentProductEntity
+import woowacourse.shopping.data.local.entity.RecentProductEntity
 
 @Dao
 interface RecentProductDao {
