+++ app/src/main/java/woowacourse/shopping/data/remote/NetworkClient.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.network
+package woowacourse.shopping.data.remote
 
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
