+++ app/src/main/java/woowacourse/shopping/data/repository/CartRepository.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.repository
+package woowacourse.shopping.data.repository
 
 import woowacourse.shopping.model.Cart
 import woowacourse.shopping.model.CartItem
