+++ app/src/main/java/woowacourse/shopping/di/AppContainer.kt
@@ -2,14 +2,14 @@ package woowacourse.shopping.di
 
 import android.content.Context
 import androidx.room.Room
-import woowacourse.shopping.local.Database
-import woowacourse.shopping.network.NetworkClient
-import woowacourse.shopping.repository.CartRepository
-import woowacourse.shopping.repository.ProductRepository
-import woowacourse.shopping.repository.RecentProductRepository
-import woowacourse.shopping.repository.network.NetworkProductRepository
-import woowacourse.shopping.repository.room.RoomCartRepository
-import woowacourse.shopping.repository.room.RoomRecentProductRepository
+import woowacourse.shopping.data.local.Database
+import woowacourse.shopping.data.remote.NetworkClient
+import woowacourse.shopping.data.repository.CartRepository
+import woowacourse.shopping.data.repository.ProductRepository
+import woowacourse.shopping.data.repository.RecentProductRepository
+import woowacourse.shopping.data.repository.remote.NetworkProductRepository
+import woowacourse.shopping.data.repository.room.RoomCartRepository
+import woowacourse.shopping.data.repository.room.RoomRecentProductRepository
 
 object AppContainer {
     private lateinit var database: Database
