+++ app/src/main/java/woowacourse/shopping/ui/shopping/ShoppingUiState.kt
@@ -1,6 +1,7 @@
 package woowacourse.shopping.ui.shopping
 
 import woowacourse.shopping.model.Products
+import woowacourse.shopping.ui.common.model.ProductUiModel
 
 data class ShoppingUiState(
     val isLoading: Boolean = false,
